refactor: responsiveness on merch detail
